requirements.txt
kopf
kubernetes
requests
redis
google-cloud-pubsub
